Setup a proper api sourceset and a datagen sourceset that will be used to generate our resources
