+++ src/api/java/com/wildfire/api/impl/GenderArmor.java
