+++ src/main/java/com/wildfire/client/resources/GenderArmorResourceManager.java
@@ -5,14 +5,10 @@ import com.google.gson.JsonElement;
 import com.wildfire.api.IGenderArmor;
 import java.util.Collections;
 import java.util.Map;
-import java.util.Optional;
-import net.minecraft.core.component.DataComponents;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
 import net.minecraft.util.profiling.ProfilerFiller;
-import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.Unmodifiable;
